Api Abstraction
Created a BotApi struct for holding function pointer that will be
called during certain events. With this, a simple bot can be
abstracted out away from all the network activity.

Also further abstracted a Connection struct to replace the simple
socket integer in order to support multiple sockets easier in the
future.