No struct typedefs
Creating typedefs for anonymous structs was going to create a problem
when attempting to forward declare a struct for the relationship
between the BotApi struct and the Bot struct. Since the BotApi creates
function pointers with references back to the calling Bot struct, and
the Bot struct contains pointers to the BotApi struct, a forward
declaration is REQUIRED, but not possible using anonymous structs. So
all typedefs to structs are now gone and will not be making a return.