No more ugly dispose
All of the dispose functions were previously macros, and they were
starting to get pretty ugly. The whole philosphy behind making them
macros was to set the pointer to NULL automatically, but I think it's
best to just get rid of them leave the responsibilty of setting things
to NULL to the consumer. The dispose calls are now just really an
extension of free(), to make sure the pointers in the structs also get
get freed.

I also started the setup for keeping track of the other players in the
room; going to use a linked list. This will help with the worry about
"who disposes of the player structs?", as the room struct will
automatically allocated free the players as they join and leave the
room.