Multiple usernames
Since the goal of the afk bot was to wait for /time to reach the
minimum time for chat restrictions (like whispering) to go away, it
makes sense to do them in bulk, so the x-arg for the afkbot now takes
in a file of a list of usernames. This will allow multiple accounts to
be timed up at the same time.