Control panel WIP
A control panel will allow commands to be sent to bots the host
machine. This should be easier than having to be in game.

In preparation, several of the socket functions found in byte_stream.c
have been moved to the connection.c file. This not only makes more
sense, but also allows the same code to be used by the panel.

A skeleton of the panel has been included in ./panel/, written in
python.